# Make inv(::Complex) return zero when input has Inf magnitude