Fix #23134.