@stevengj 2019-05-17
What is the performance impact of this?