@giordano 2019-05-17
Before the PR:
```julia
julia> v64 = randn(ComplexF64, 100000);

julia> @btime inv.($v64);
  1.393 ms (2 allocations: 1.53 MiB)

julia> v32 = randn(ComplexF32, 100000);

julia> @btime inv.($v32);
  426.785 μs (2 allocations: 781.33 KiB)

julia> vbig = big.(randn(ComplexF64, 1000));

julia> @btime inv.($vbig);
  1.057 ms (13001 allocations: 695.44 KiB)
```

After the PR:
```julia
julia> v64 = randn(ComplexF64, 100000);

julia> @btime inv.($v64);
  1.522 ms (2 allocations: 1.53 MiB)

julia> v32 = randn(ComplexF32, 100000);

julia> @btime inv.($v32);
  1.601 ms (2 allocations: 781.33 KiB)

julia> vbig = big.(randn(ComplexF64, 1000));

julia> @btime inv.($vbig);
  1.064 ms (13001 allocations: 695.44 KiB)
```

So, `ComplexF64` has a ~10% penalty, `Complex{BigFloat}` is basically unchanged, `ComplexF32` is much worse.  Note that for `ComplexF32` I changed the operation performed (before it was `conj(widen(z))/abs2(widen(z))`, in the PR it is `inv(widen(z))`).

Note also that as it is now, `ComplexF32` and `Complex{BigFloat}` give inconsistent results compared to `ComplexF64`:
```julia
julia> inv(complex(Inf32))
NaN32 - 0.0f0im

julia> inv(complex(big(Inf)))
NaN - 0.0im

julia> inv(complex(Inf))
0.0 - 0.0im
```