@stevengj 2019-05-17
With a big array, you are measuring cache and allocation time as well as computation.