@stevengj 2019-05-17
I tried
```
julia> a = rand(ComplexF64, 1000);

julia> f(x) = sum(inv, x)
f (generic function with 1 method)

julia> g(x) = sum(newinv, x)
g (generic function with 1 method)

julia> @btime f($a);
  14.100 μs (0 allocations: 0 bytes)

julia> @btime g($a);
  2.306 μs (0 allocations: 0 bytes)
```
so the new version is 7x faster?  Seems weird.