@giordano 2019-05-17
I don't know if that's the cause, but `conj(z)/abs2(z)` calls three times `real(z)` and other three times `imag(z)`, the new version calls `reim` only once