@stevengj 2019-05-17
It's maybe a change in whether it inlines?