@stevengj 2019-05-17
If I do
```jl
function newinv2(z::Complex)
    c, d = reim(z)
    complex(c, -d)/(c * c + d * d)
end
```
then `g2(x) = sum(newinv2, x)` gives
```jl
julia> @btime g2($a);
  1.125 μs (0 allocations: 0 bytes)
```
which is 2x faster than the version with the `isinf` checks.