@stevengj 2019-05-17
Oh, I just realized that I'm comparing the wrong version — I need the `inv(w::ComplexF64)`  version.