@chethega 2019-05-18
For Complex{Float32}, my machine counts a slowdown of 2.3 -> 8.5 cycles in batch / simd, and 6 -> 11 cycles for sequential processing. With `Complex{Float64}`, I count 7->8 and 7->10.5 cycles (on 2ghz broadwell). That's pretty expensive for a very rare corner case.
```
julia> using BenchmarkTools
julia> function _inv_new(z::Complex)
       c, d = reim(z)
       (isinf(c) | isinf(d)) && return complex(copysign(zero(c), c), flipsign(-zero(d), d))
       complex(c, -d)/(c * c + d * d)
       end
julia> _inv_old(z)=conj(z)/abs2(z);
julia> r=rand(Complex{Float32}, 1000); b=copy(r);
julia> @btime broadcast!($_inv_old, $b, $r); @btime broadcast!($_inv_new, $b, $r); @btime _inv_old($r[1]); @btime _inv_new($r[1]);
  1.143 μs (0 allocations: 0 bytes)
  4.268 μs (0 allocations: 0 bytes)
  3.075 ns (0 allocations: 0 bytes)
  5.506 ns (0 allocations: 0 bytes)
julia> r=rand(Complex{Float64}, 1000); b=copy(r);

julia> @btime broadcast!($_inv_old, $b, $r); @btime broadcast!($_inv_new, $b, $r); @btime _inv_old($r[1]); @btime _inv_new($r[1]);
  3.516 μs (0 allocations: 0 bytes)
  3.992 μs (0 allocations: 0 bytes)
  3.516 ns (0 allocations: 0 bytes)
  5.262 ns (0 allocations: 0 bytes)
```