@giordano 2019-05-18
I run your benchmark on my machine and I get very different results
```julia
julia> @btime broadcast!($_inv_old, $b, $r); @btime broadcast!($_inv_new, $b, $r); @btime _inv_old($r[1]); @btime _inv_new($r[1]);
  1.095 μs (0 allocations: 0 bytes)
  2.793 μs (0 allocations: 0 bytes)
  3.811 ns (0 allocations: 0 bytes)
  3.287 ns (0 allocations: 0 bytes)

julia> r=rand(Complex{Float64}, 1000); b=copy(r);

julia> @btime broadcast!($_inv_old, $b, $r); @btime broadcast!($_inv_new, $b, $r); @btime _inv_old($r[1]); @btime _inv_new($r[1]);
  3.810 μs (0 allocations: 0 bytes)
  3.811 μs (0 allocations: 0 bytes)
  3.812 ns (0 allocations: 0 bytes)
  3.815 ns (0 allocations: 0 bytes)
```
For `Float64` there is basically no change, for `Float32` the scalar benchmark seems to favour the new version