@stevengj 2019-05-20
@chethega, you made the same mistake I did: the `conj(z)/abs2(z)` is not the method that is called for `ComplexF32` or `ComplexF64`.   See the `inv(w::ComplexF64)`, which has a bunch of extra checks to handle overflow/underflow already.