@StefanKarpinski 2019-06-04
@stevengj, please merge if this looks good to you.